dsfgdfg

+++ personws/src/main/java/ru/company/services/personws/type/TUserListResponse.java
@@ -24,7 +24,7 @@ public class TUserListResponse {
     List<TUser> users;
 
     @XmlElement
-    Integer totalCount;
+    Long totalCount;
 
     public TUserListResponse() {
         this.users = new ArrayList<TUser>();
